+++ src/Umbraco.Abstractions/IO/IOHelper.cs
@@ -255,17 +255,6 @@ namespace Umbraco.Core.IO
             _rootDir = rootPath;
         }
 
-        /// <summary>
-        /// Check to see if filename passed has any special chars in it and strips them to create a safe filename.  Used to overcome an issue when Umbraco is used in IE in an intranet environment.
-        /// </summary>
-        /// <param name="filePath">The filename passed to the file handler from the upload field.</param>
-        /// <returns>A safe filename without any path specific chars.</returns>
-        public string SafeFileName(string filePath)
-        {
-            // use string extensions
-            return filePath.ToSafeFileName();
-        }
-
         public void EnsurePathExists(string path)
         {
             var absolutePath = MapPath(path);
